Crate ibc_core_handler

Exposes IBC handler entry points for an integrated IBC core modules. These entry points are responsible for processing incoming IBC messages, performing validation, and execution logics by invoking the appropriate module handler.

When processing a given message M, if any method in this library returns an error, the runtime is expected to rollback all state modifications made to the context (e.g. ExecutionContext) while processing M. If the transaction containing M consists of multiple messages, then typically the state modifications from all messages are expected to be rolled back as well.

Modules

entrypoint

types

Re-export IBC handler types from ibc-core-handler-types crate.